time:int->long
Former-commit-id: 8f6857546dde01a232b2047a147ee7aa3132b58b

+++ back-end-projects/Explorer/src/main/java/com/github/ontio/service/impl/BlockServiceImpl.java
@@ -127,10 +127,10 @@ public class BlockServiceImpl implements IBlockService {
     }
 
     @Override
-    public Result blockCountInTwoWeeks(int time) {
+    public Result blockCountInTwoWeeks(long time) {
 
         List<Map> twoWeeksBlockCountList = new ArrayList<>();
-        int zeroTime = time / 86400 * 86400;
+        long zeroTime = time / 86400 * 86400;
         try {
             for (int i = 0; i < 14; i++) {
                 if (i == 0) {
@@ -140,8 +140,8 @@ public class BlockServiceImpl implements IBlockService {
                     oneDayCount.put("count", count);
                     twoWeeksBlockCountList.add(oneDayCount);
                 } else {
-                    int dayEndTime = zeroTime - (i - 1) * 86400;
-                    int dayStartTime = dayEndTime - 86400;
+                    long dayEndTime = zeroTime - (i - 1) * 86400;
+                    long dayStartTime = dayEndTime - 86400;
                     int count = blockMapper.selectBlockCountInOneDay(dayStartTime, dayEndTime);
                     Map<String, Object> oneDayCount = new HashMap<>();
                     oneDayCount.put("day", dayStartTime);
